+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+### Added
+- `delay-startup` option to the `blink` command.
 
 ## [0.3.0] - 2021-02-13
 ### Added